refactor: Created interface and implementation of it
