+++ src/main/java/uk/gov/hmcts/reform/cwrdapi/service/impl/CaseWorkerStaticValueRepositoryAccessorImpl.java
@@ -1,20 +1,19 @@
-package uk.gov.hmcts.reform.cwrdapi.util;
+package uk.gov.hmcts.reform.cwrdapi.service.impl;
 
-import lombok.Getter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.support.SimpleJpaRepository;
 import org.springframework.stereotype.Component;
 
 import uk.gov.hmcts.reform.cwrdapi.domain.RoleType;
 import uk.gov.hmcts.reform.cwrdapi.domain.UserType;
+import uk.gov.hmcts.reform.cwrdapi.service.CaseWorkerStaticValueRepositoryAccessor;
 
 import java.util.Collections;
 import java.util.List;
 import javax.annotation.PostConstruct;
 
 @Component
-@Getter
-public class CaseWorkerStaticValueRepositoryAccessor {
+public class CaseWorkerStaticValueRepositoryAccessorImpl implements CaseWorkerStaticValueRepositoryAccessor {
     private List<RoleType> roleTypes;
 
     private List<UserType> userTypes;
@@ -29,5 +28,13 @@ public class CaseWorkerStaticValueRepositoryAccessor {
     public void initialize() {
         roleTypes = Collections.unmodifiableList(roleTypeRepository.findAll());
         userTypes = Collections.unmodifiableList(userTypeRepository.findAll());
+    }
+
+    public List<RoleType> getRoleTypes() {
+        return roleTypes;
+    }
+
+    public List<UserType> getUserTypes() {
+        return userTypes;
     }
 }
